cleanup

+++ src/java/org/apache/cassandra/db/ReadCommand.java
@@ -199,7 +199,7 @@ public abstract class ReadCommand extends AbstractReadQuery
             mutations.forEach(this::augment);
         }
 
-        int nowInSec();
+        long nowInSec();
         void close();
     }
